Create MyFirstProgram
This is my first program in python, it collects information form user input and compares it to recommended savings and determines if the user's savings are on track or not. It uses functions to collect the information from the user and then adds the values to lists, the lists are then returned from the functions and stripped out to create variables for input into other functions. 

I have no clue if this method of returning multiple values from a function is effective or efficient, but it was fun thinking of a way to be able to return more than a single value.

The purpose of writing this program for me was I wanted to practice creating and returning output from functions. It was effective for that and as a bonus I got to practice using lists and appending them and indexing values in a list.